Replace some shams
+ Added task to util. There wasn't a task corrollary in central.
+ Added promise middleware to util. I wanted to add it to devtools, but
it still relies on promise as opposed to defer, because we haven't
converted those modules yet
+ Created util/sources for isJavaScript. I considered copying
debugger/utils into devtools, but ran into two issues (L10N, and
exports).

+ Move the middleware
+ Move task
