+++ public/js/util/redux/middleware/history.js
