+++ public/js/util/task.js
@@ -6,8 +6,6 @@
 
 "use strict";
 
-// const promise = require("devtools-sham/sham/promise");
-
 /**
  * This object provides the public module functions.
  */
